refactor(hypercore): Don't expose testing
